Break out beaker and fixture tasks into separate files
This makes them easier to consume on their own without pulling in
everything. This also moves around the tests for those parts, and
adds smoke tests to make sure Rakefiles still load and process
correctly.
